(#79) refactor: 인증 페이지 컴포넌트 구조 개선 및 성능 최적화
