+++ package.json
@@ -11,6 +11,7 @@
     "prepare": "lefthook install"
   },
   "dependencies": {
+    "@hookform/resolvers": "5.1.1",
     "@tanstack/react-query": "5.81.5",
     "browser-image-compression": "2.0.2",
     "class-variance-authority": "0.7.1",
